extract export-lanes functionality to a new file
